+++ Neolution.Extensions.Identity.Abstractions/ISignInManager.cs
@@ -25,7 +25,7 @@
         /// Used to ensure that a user is allowed to sign in.
         /// </summary>
         /// <param name="user">The user</param>
-        /// <returns>Null if the user should be allowed to sign in, otherwise the SignInResponse why they should be denied.</returns>
-        Task<SignInResponse?> PreSignInCheckAsync(TUser user);
+        /// <returns>The task object representing the asynchronous operation containing the <see name="PreSignInResponse"/> for the pre sign-in check.</returns>
+        Task<PreSignInResponse> PreSignInCheckAsync(TUser user);
     }
 }